src/rice_disease_prediction/requirements.txt
# local package
-e .
pandas
numpy
matplotlib
seaborn
plotly
statsmodels
pmdarima
gunicorn
flask
xgboost
Flask-SQLAlchemy
Flask
streamlit
# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
torch
torchvision
pillow
